requirements.txt
streamlit
openai
faiss-cpu
pymupdf
numpy
tqdm
python-dotenv
tiktoken
beautifulsoup4
markdown2